warnings

+++ Frameworks/OakFoundation/src/NSString Additions.h
@@ -1,4 +1,4 @@
 @interface NSString (Additions)
-+ (NSString*)stringWithUTF8String:(char const*)aString length:(unsigned)aLength;
++ (NSString*)stringWithUTF8String:(char const*)aString length:(NSUInteger)aLength;
 + (NSString*)stringWithCxxString:(std::string const&)aString;
 @end
